+++ tests/testthat/test-prism_data_get.R
@@ -16,14 +16,14 @@ exp <- c(
   "Apr 30-year normals - 4km resolution - Maximum vapor pressure deficit"
 )
 
-test_that("prism_data_get_name() works.", {
-  expect_identical(prism_data_get_name(tst_files), exp)
+test_that("pd_get_name() works.", {
+  expect_identical(pd_get_name(tst_files), exp)
   expect_identical(expect_warning(prism_md(tst_files)), exp)
 })
 
 exp <- c("1967-06-15", "2020-04-01", "", "2019-01-01", "1967-10-01", "")
 
-test_that("prism_data_get_date() works.", {
-  expect_identical(prism_data_get_date(tst_files), exp)
+test_that("pd_get_date() works.", {
+  expect_identical(pd_get_date(tst_files), exp)
   expect_identical(expect_warning(prism_md(tst_files, TRUE)), exp)
 })